refactor: split resolver module
